+++ src/utils/render.js
@@ -63,3 +63,4 @@ export const remove = (component) => {
   component.element.remove();
   component.removeElement();
 };
+
