+++ src/view/abstract-view.js
@@ -26,3 +26,4 @@ export default class AbstractView {
     this.#element = null;
   }
 }
+
