+++ src/view/header-filters-view/header-filters-view.js
@@ -1,0 +1,9 @@
+import AbstractView from '../abstract-view';
+import { createHeaderFiltersTemplate } from './header-filters-view.tlp.js';
+
+export default class HeaderFiltersView extends AbstractView {
+  get template() {
+    return createHeaderFiltersTemplate();
+  }
+}
+
