+++ src/view/header-filters-view/header-filters-view.tlp.js
@@ -1,6 +1,4 @@
-import AbstractView from './abstract-view';
-
-const createHeaderFiltersTemplate = () =>
+export const createHeaderFiltersTemplate = () =>
   `<form class="trip-filters" action="#" method="get">
       <div class="trip-filters__filter">
       <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
@@ -20,8 +18,3 @@ const createHeaderFiltersTemplate = () =>
       <button class="visually-hidden" type="submit">Accept filter</button>
   </form>`;
 
-export default class HeaderFiltersView extends AbstractView {
-  get template() {
-    return createHeaderFiltersTemplate();
-  }
-}
