+++ src/view/header-info-view/header-info-view.js
@@ -1,0 +1,9 @@
+import AbstractView from '../abstract-view';
+import { createHeaderInfoTemplate } from './header-info-view.tlp';
+
+export default class HeaderInfoView extends AbstractView {
+  get template() {
+    return createHeaderInfoTemplate();
+  }
+}
+
