--- src/view/header-menu-view.js
@@ -1,13 +1,0 @@
-import AbstractView from './abstract-view';
-
-const createHeaderMenuTemplate = () =>
-  `<nav class="trip-controls__trip-tabs  trip-tabs">
-      <a class="trip-tabs__btn" href="#">Table</a>
-      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
-  </nav>`;
-
-export default class HeaderMenuView extends AbstractView {
-  get template() {
-    return createHeaderMenuTemplate();
-  }
-}
