+++ src/view/header-menu-view/header-menu-view.js
@@ -1,0 +1,10 @@
+import AbstractView from '../abstract-view';
+import { createHeaderMenuTemplate } from './header-menu-view.tlp';
+
+export default class HeaderMenuView extends AbstractView {
+  get template() {
+    return createHeaderMenuTemplate();
+  }
+}
+
+
