+++ src/view/header-menu-view/header-menu-view.tlp.js
@@ -1,0 +1,6 @@
+export const createHeaderMenuTemplate = () =>
+  `<nav class="trip-controls__trip-tabs  trip-tabs">
+      <a class="trip-tabs__btn" href="#">Table</a>
+      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
+  </nav>`;
+
