+++ src/view/main-form-view/main-form-view.js
@@ -1,0 +1,35 @@
+import AbstractView from '../abstract-view';
+import { createMainFormTemplate } from './main-form-view.tlp';
+
+export default class MainFormView extends AbstractView {
+  #point = null;
+
+  constructor(point) {
+    super();
+    this.#point = point;
+  }
+
+  get template() {
+    return createMainFormTemplate(this.#point);
+  }
+
+  setFormSubmitHandler = (callback) => {
+    this._callback.formSubmit = callback;
+    this.element.querySelector('form').addEventListener('submit', this.#formSubmitHandler);
+  }
+
+  #formSubmitHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.formSubmit();
+  }
+
+  setEditClickHandler = (callback) => {
+    this._callback.editClick = callback;
+    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#editClickHandler);
+  }
+
+  #editClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.editClick();
+  }
+}
