+++ src/view/main-sort-form-view/main-sort-form-view.js
@@ -1,0 +1,8 @@
+import AbstractView from '../abstract-view';
+import { createMainSortFormTemplate } from './main-sort-form-view.tlp';
+
+export default class MainSortFormView extends AbstractView {
+  get template() {
+    return createMainSortFormTemplate();
+  }
+}
