+++ src/view/main-trip-list-view/main-trip-list-view.js
@@ -1,9 +1,10 @@
-import AbstractView from './abstract-view';
-
-const createMainTripListTemplate = () => '<ul class="trip-events__list"></ul>';
+import AbstractView from '../abstract-view';
+import { createMainTripListTemplate } from './main-trip-list-view.tlp';
 
 export default class MainTripListView extends AbstractView {
   get template() {
     return createMainTripListTemplate();
   }
 }
+
+
