+++ src/view/main-trip-list-view/main-trip-list-view.tlp.js
@@ -1,0 +1,2 @@
+export const createMainTripListTemplate = () => '<ul class="trip-events__list"></ul>';
+
