+++ src/view/main-trip-no-point-view/main-trip-no-point-view.js
@@ -1,0 +1,8 @@
+import AbstractView from '../abstract-view';
+import { createNoPointTemplate } from './main-trip-no-point-view.tlp';
+
+export default class NoPointView extends AbstractView {
+  get template() {
+    return createNoPointTemplate();
+  }
+}
