+++ src/view/main-trip-no-point-view/main-trip-no-point-view.tlp.js
@@ -1,0 +1,5 @@
+export const createNoPointTemplate = () =>
+  `<p class="trip-events__msg">
+Click New Event to create your first point
+</p>`;
+
