+++ src/view/main-trip-point-view/main-trip-point-view.js
@@ -1,0 +1,25 @@
+import AbstractView from '../abstract-view';
+import { createMainTripPointTemplate } from './main-trip-point-view.tlp';
+
+export default class MainTripPointView extends AbstractView {
+  #point = null;
+
+  constructor(point) {
+    super();
+    this.#point = point;
+  }
+
+  get template() {
+    return createMainTripPointTemplate(this.#point);
+  }
+
+  setEditClickHandler = (callback) => {
+    this._callback.editClick = callback;
+    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#editClickHandler);
+  }
+
+  #editClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.editClick();
+  }
+}
