+++ src/view/main-trip-point-view/main-trip-point-view.tlp.js
@@ -1,7 +1,6 @@
 import dayjs from 'dayjs';
-import AbstractView from './abstract-view';
 
-const createMainTripPointTemplate = (point) => {
+export const createMainTripPointTemplate = (point) => {
   const {
     basePrice,
     dateFrom,
@@ -50,25 +49,3 @@ const createMainTripPointTemplate = (point) => {
   </li>`;
 };
 
-export default class MainTripPointView extends AbstractView {
-  #point = null;
-
-  constructor(point) {
-    super();
-    this.#point = point;
-  }
-
-  get template() {
-    return createMainTripPointTemplate(this.#point);
-  }
-
-  setEditClickHandler = (callback) => {
-    this._callback.editClick = callback;
-    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#editClickHandler);
-  }
-
-  #editClickHandler = (evt) => {
-    evt.preventDefault();
-    this._callback.editClick();
-  }
-}
